ISIC exports

+++ app/models/isic_export.rb
@@ -11,10 +11,6 @@ class IsicExport < ActiveRecord::Base
     Member.includes(:current_card, :club).where(:id => self.members, :enabled => true)
   end
 
-  before_create do
-    self.members = Member.find_all_for_isic_export.map(&:id)
-  end
-
   # Generate the export files
   def generate
     members = self.full_members
@@ -29,12 +25,14 @@ class IsicExport < ActiveRecord::Base
       end
 
       if card.isic_status == 'request'
-        member.current_card.isic_status = 'requested'
+        card.isic_status = 'requested'
       elsif card.isic_status == 'revalidate'
         # Don't do anything
       else
         raise "Unable to handle isic_status #{member.current_card.isic_status}"
       end
+
+      # TODO: don't handle cards that were not in the original request
 
       card.isic_exported = 1
       card.save!
@@ -53,13 +51,14 @@ class IsicExport < ActiveRecord::Base
     sheet = book.create_worksheet :name => "Gegevens"
     sheet.row(0).concat ["School", "Kring", "Voornaam", "Familienaam",
         "Geboortedatum", "E-mailadres", "Thuisadres", "FK-nummer", "Foto",
-        "ISIC Nieuwsbrief", "Kaart opsturen"]
+        "ISIC status", "ISIC nieuwsbrief", "Kaart opsturen"]
 
     members.each_with_index do |member, i|
       sheet.row(i+1).concat ["UGent", member.club.internal_name,
           member.first_name, member.last_name, member.date_of_birth, member.email,
           member.home_address.sub("\r\n", "\n"), member.current_card.number,
-          "#{member.id}.jpg", member.isic_newsletter, member.isic_mail_card]
+          "#{member.id}.jpg", member.current_card.isic_status, member.isic_newsletter,
+          member.isic_mail_card]
     end
     book.write(filename)
     assign_file_and_unlink(:data, filename)
@@ -81,6 +80,8 @@ class IsicExport < ActiveRecord::Base
 
 private
   def assign_file_and_unlink(property, filename)
+    return unless File.exists? filename
+
     File.open(filename) do |file|
       self.send("#{property}=", file)
     end
